+++ querydsl-jpa/src/main/java/com/querydsl/jpa/hibernate/HibernateDeleteClause.java
@@ -13,14 +13,6 @@
  */
 package com.querydsl.jpa.hibernate;
 
-import java.util.HashMap;
-import java.util.Map;
-
-import org.hibernate.LockMode;
-import org.hibernate.query.Query;
-import org.hibernate.Session;
-import org.hibernate.StatelessSession;
-
 import com.querydsl.core.JoinType;
 import com.querydsl.core.dml.DeleteClause;
 import com.querydsl.core.support.QueryMixin;
@@ -31,6 +23,13 @@ import com.querydsl.jpa.HQLTemplates;
 import com.querydsl.jpa.JPAQueryMixin;
 import com.querydsl.jpa.JPQLSerializer;
 import com.querydsl.jpa.JPQLTemplates;
+import org.hibernate.LockMode;
+import org.hibernate.Session;
+import org.hibernate.StatelessSession;
+import org.hibernate.query.Query;
+
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * DeleteClause implementation for Hibernate
@@ -75,8 +74,7 @@ public class HibernateDeleteClause implements DeleteClause<HibernateDeleteClause
         for (Map.Entry<Path<?>, LockMode> entry : lockModes.entrySet()) {
             query.setLockMode(entry.getKey().toString(), entry.getValue());
         }
-        HibernateUtil.setConstants(query, serializer.getConstantToNamedLabel(), serializer.getConstantToNumberedLabel(),
-                queryMixin.getMetadata().getParams());
+        HibernateUtil.setConstants(query, serializer.getConstants(), queryMixin.getMetadata().getParams());
         return query.executeUpdate();
     }