+++ querydsl-jpa/src/main/java/com/querydsl/jpa/impl/JPADeleteClause.java
@@ -13,8 +13,6 @@
  */
 package com.querydsl.jpa.impl;
 
-import java.util.Map;
-
 import org.jetbrains.annotations.Nullable;
 import javax.persistence.EntityManager;
 import javax.persistence.LockModeType;
@@ -60,13 +58,12 @@ public class JPADeleteClause implements DeleteClause<JPADeleteClause> {
     public long execute() {
         JPQLSerializer serializer = new JPQLSerializer(templates, entityManager);
         serializer.serializeForDelete(queryMixin.getMetadata());
-        Map<Object,String> constants = serializer.getConstantToAllLabels();
 
         Query query = entityManager.createQuery(serializer.toString());
         if (lockMode != null) {
             query.setLockMode(lockMode);
         }
-        JPAUtil.setConstants(query, constants, queryMixin.getMetadata().getParams());
+        JPAUtil.setConstants(query, serializer.getConstants(), queryMixin.getMetadata().getParams());
         return query.executeUpdate();
     }